Update to val_to_percent function in index.js
The current val_to_percent function returns percentages that are skewed more towards the lower end of the range because you are using a hex encoding, which assigns more probability to numbers 0-5 that correspond to the hex letters a-f. In practice, when I ran that function 100,000 with random parameters I would get a distribution of 63-37%.  The proposed changes make it a 50-50% distribution.